Model-average dose-response curves

Computing stacking weights.

getStackingWeights(modelList, nSplits = 2)

Arguments

modelList

list of models of class drc

nSplits

integer. The Stacking Weights are based on V-fold cross-validation. The nSplits argument sets the number V of data splits used in the cross validation.

Value

numeric

Details

The aim to provide an R package calculating the benchmark dose (BMD) and the lower limit of the corresponding 95% confidence interval (BMDL) for continuous and quantal dose-response data for a range of dose-response models based on the available definitions of the benchmark dose concepts.

Details on the implemented definitions and methods can be found in Crump (2002)

Author

Jens Riis Baalkilde

Examples

library(bmd)
library(drc)
library(drcData)

# fit models to aconiazide data
aconiazide.LL.3 <- drm(weightChange ~ dose,data = aconiazide,fct = LL.3())
aconiazide.LN.3 <- drm(weightChange ~ dose,data = aconiazide,fct = LN.3())
aconiazide.W1.3 <- drm(weightChange ~ dose,data= aconiazide,fct = W1.3())
aconiazide.W2.3 <- drm(weightChange ~ dose,data= aconiazide,fct = W2.3())

getStackingWeights(modelList = list(aconiazide.LL.3, aconiazide.LN.3,
                                  aconiazide.W1.3, aconiazide.W2.3))
#> [1] 2.221066e-07 1.741480e-07 5.000002e-01 4.999994e-01
getStackingWeights(modelList = list(aconiazide.LL.3, aconiazide.LN.3,
                                  aconiazide.W1.3, aconiazide.W2.3),
                  nSplits = 7)
#> [1] 3.155175e-08 6.086256e-10 7.142857e-01 2.857143e-01